* 'master' of https://github.com/SoulXHades/FindMyFirstHome:
  aa

+++ app/src/main/java/com/example/findmyfirsthome/Boundary/ProfileUI2.java
@@ -182,7 +182,7 @@ public class ProfileUI2 extends AppCompatActivity implements View.OnFocusChangeL
                 if (value.isEmpty() == true) {
                     x.setText("0");
                 }
-                else if(Integer.parseInt(x.getText().toString())<0){
+                else if(Double.parseDouble(x.getText().toString())<0){
                     Toast.makeText(getApplicationContext(), "Commitments cannot be negative!", Toast.LENGTH_SHORT).show();
                     x.setText("0");
                 }
